// models/site-meta.ts
export type SiteMeta = {
    id: string;
    site_id: string;
    slug: string;
    name: string;
    tel: string;
    email: string;
    postal_code: string;
    address: string;
    building: string;
    access: string;
    description: string | null;
    background_image: string | null;
    avatar: string | null;
    created_at: string;
    updated_at: string;
};